git script: correct erased branch substitution
During the extraction of a job configuration, the substitution of variables
is done before git script starts looking for existing branches in the git repo.
If the template has already a branch or a commit specified, it is automatically
erased without checking.

This commit checks if such a treeish exists and avoid erasing it.