autojob: add noscm python script
Up to now, existing scripts copy jenkins jobs according to their SCM.
So a jenkins job has to have a SCM to be copied.

noscm python script allows copying a template jenkins job without any
SCM.  This script usage is the same as the others.

In order to keep main.py happy, the list_branches function returns a
single dummy branch, which contains the name of the template imported.
This branch won't be used anyway, it only allows noscm to create one
new job.